refactor(desktop): sink daemon probe helpers from uc-tauri
Move the GUI-framework agnostic daemon probe helpers
(probe_daemon_health / classify_health_response /
load_daemon_connection_info / terminate_incompatible_daemon_from_pid_file
/ replay_lifecycle_ready_after_respawn plus all the timing constants)
from uc-tauri/src/bootstrap/run.rs into uc-desktop/src/daemon_probe.rs.
Future shells (uc-macos-native etc.) writing their own spawn-supervise
loops can now reuse these without re-implementing them.

classify_health_response now takes expected_package_version as a
parameter rather than baking in env!("CARGO_PKG_VERSION") — the
desktop crate has no opinion on which shell version the daemon
should match. uc-tauri pins it to its own version via a constant.

uc-tauri/src/bootstrap/run.rs keeps only the Tauri-specific glue:
spawn_daemon_process (uses tauri-plugin-shell's sidecar API),
bootstrap_daemon_connection (consumes AppHandle), and supervise_daemon
(the respawn loop bound to AppHandle). 412 lines down to 256.

Verified cargo tree -p uc-desktop -e normal still shows zero
tauri / tauri-plugin edges.
