+++ src-tauri/crates/uc-desktop/Cargo.toml
@@ -32,6 +32,7 @@ blake3 = "1"
 tracing = "0.1"
 futures-util = "0.3"
 url = "2"
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls", "rustls-tls-webpki-roots"] }
 
 [dev-dependencies]
 uc-daemon-local = { path = "../uc-daemon-local", features = ["test-helpers"] }
