+++ src-tauri/crates/uc-desktop/src/lib.rs
@@ -15,6 +15,7 @@ pub use uc_daemon_contract::DAEMON_API_REVISION;
 
 pub mod background;
 pub mod daemon;
+pub mod daemon_probe;
 pub mod runtime;
 pub mod sidecar;
